+++ client/.env
@@ -1,0 +1,2 @@
+REACT_APP_AUTH0_DOMAIN=dev-dist0ul4.au.auth0.com
+REACT_APP_AUTH0_CLIENT_ID=DhirGPL0sRdQbAST2cB0bmPo6cbyE0Ym
