--- client/.env.local
@@ -1,3 +1,0 @@
-REACT_APP_AUTH0_DOMAIN=dev-dist0ul4.au.auth0.com
-REACT_APP_AUTH0_CLIENT_ID=DhirGPL0sRdQbAST2cB0bmPo6cbyE0Ym
-REACT_APP_GOOGLE_MAPS_API_KEY=AIzaSyBcDAsOxlL1G1FNDii14G2DjiO1olZeT5Y
