+++ client/src/index.js
@@ -6,7 +6,6 @@ import registerServiceWorker from "./registerServiceWorker";
 import { BrowserRouter as Router } from "react-router-dom";
 import Auth0ProviderWithHistory from "./auth/auth0-provider-with-history"
 import { ExternalLink } from 'react-external-link';
-import { GoogleMap } from '@react-google-maps/api';
 
 ReactDOM.render(
     <Router>
